+++ lib/screens/Home/home.dart
@@ -1,7 +1,8 @@
 import "package:flutter/material.dart";
 import "package:flutter_svg/flutter_svg.dart";
 import "package:google_fonts/google_fonts.dart";
-import "package:recychamp/ui/bottom_app_bar.dart";
+import "package:recychamp/screens/Challenges/challenges.dart";
+import "package:recychamp/screens/Dashboard/dashboard.dart";
 import "package:recychamp/ui/home_three_row_button.dart";
 import "package:recychamp/ui/latest_challenges_card.dart";
 
@@ -13,244 +14,82 @@ class Home extends StatefulWidget {
 }
 
 class _HomeState extends State<Home> {
+  // * Dashboard is selected as home screen
+  int _selectedIndex = 2;
+
+  static const List<Widget> _widgetOptions = <Widget>[
+    // * Add appropriate screens as commented below
+    Challenges(), // * Challenges
+    Challenges(), // * Search
+    Dashboard(), // * Dashboard
+    Challenges(), // * Community
+    Challenges(), // * Notifications
+  ];
+
+  void _onItemTapped(int index) {
+    setState(() {
+      _selectedIndex = index;
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
-    // * Client device data (screen height & width) for responsiveness
-    var deviceData = MediaQuery.of(context);
-
-    print(deviceData.size.height);
     return Scaffold(
-      bottomNavigationBar: const CustomBottomAppBar(),
-      resizeToAvoidBottomInset: false,
-      backgroundColor: const Color(0xff75A488),
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          const SizedBox(
-            height: 50,
-          ),
-          Container(
-            margin:
-                EdgeInsets.symmetric(horizontal: deviceData.size.width * 0.07),
-            // * Headline row with settings button
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                // * Headline left column
-                Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    // todo greeting must change according to the time of the day
-                    Text(
-                      "Good Morning",
-                      style: GoogleFonts.almarai(
-                        color: Colors.white,
-                        fontSize: 25,
-                        fontWeight: FontWeight.w700,
-                      ),
-                    ),
-                    Text(
-                      "Susan Clay",
-                      style: GoogleFonts.almarai(
-                        color: Colors.white,
-                        fontSize: 16,
-                        fontWeight: FontWeight.w400,
-                      ),
-                    )
-                  ],
-                ),
-                // todo Settings button (must link to settings page)
-                SvgPicture.asset(
-                  "assets/icons/Settings.svg",
-                  height: 24,
-                  width: 24,
-                )
-              ],
-            ),
-          ),
-          const SizedBox(
-            height: 58,
-          ),
-          Expanded(
-            // * Curved section
-            child: FractionallySizedBox(
-              // heightFactor: 0.85,
-              child: Container(
-                width: double.infinity,
-                // height: double.infinity,
-                decoration: const ShapeDecoration(
-                  color: Colors.white,
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.only(
-                      topLeft: Radius.circular(42.07),
-                      topRight: Radius.circular(42.07),
-                    ),
-                  ),
-                ),
-                child: Stack(
-                  clipBehavior: Clip.none,
-                  alignment: Alignment.center,
-                  children: [
-                    // * Profile picture frame
-                    Positioned(
-                      height: 84.15,
-                      top: -84.15 / 2,
-                      child: Container(
-                        width: 84.15,
-                        height: 84.15,
-                        decoration: ShapeDecoration(
-                          gradient: const LinearGradient(
-                            begin: Alignment(0.00, 1.00),
-                            end: Alignment(0, -1),
-                            colors: [
-                              Color(0xFFFEAA42),
-                              Color(0xFFFBA33F),
-                              Color(0xFFF59838),
-                              Color(0xFFF29135),
-                              Color(0xFFF18F34),
-                              Color(0xFFF28F3E),
-                              Color(0xFFF38E5A),
-                              Color(0xFFF68D88),
-                              Color(0xFFF78C9B),
-                              Color(0xFFF08672)
-                            ],
-                          ),
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(16.83),
-                          ),
-                        ),
-                      ),
-                    ),
-                    // * remaining content
-                    Column(
-                      children: [
-                        Expanded(
-                          child: Container(
-                            margin: const EdgeInsets.only(top: 50),
-                            child: Column(
-                              children: [
-                                SvgPicture.asset(
-                                  "assets/icons/gold-medal.svg",
-                                  height: 32.08,
-                                  width: 32.08,
-                                ),
-                                Expanded(
-                                  child: Column(
-                                    children: [
-                                      SizedBox(
-                                        height: 75,
-                                        child: Text.rich(
-                                          TextSpan(
-                                            children: [
-                                              TextSpan(
-                                                text:
-                                                    "Your fresh and green comfortable ",
-                                                style: GoogleFonts.almarai(
-                                                    color:
-                                                        const Color(0xFF1E1E1E),
-                                                    fontSize: 27,
-                                                    fontWeight:
-                                                        FontWeight.w700),
-                                              ),
-                                              TextSpan(
-                                                text: "place",
-                                                style: GoogleFonts.almarai(
-                                                    color:
-                                                        const Color(0xFF75A488),
-                                                    fontSize: 27,
-                                                    fontWeight:
-                                                        FontWeight.w700),
-                                              ),
-                                            ],
-                                          ),
-                                          textAlign: TextAlign.center,
-                                        ),
-                                      ),
-                                      const SizedBox(
-                                        height: 20,
-                                      ),
-
-                                      // * 3 button row
-                                      SizedBox(
-                                        height: 113.60,
-                                        child: Container(
-                                          margin: EdgeInsets.symmetric(
-                                              horizontal:
-                                                  deviceData.size.width * 0.05),
-                                          child: const Row(
-                                            mainAxisAlignment:
-                                                MainAxisAlignment.spaceBetween,
-                                            children: [
-                                              HomeThreeRowButton(
-                                                  iconURL:
-                                                      "assets/icons/home-article.svg",
-                                                  description: "Articles"),
-                                              HomeThreeRowButton(
-                                                  iconURL:
-                                                      "assets/icons/home-cart.svg",
-                                                  description: "Visit Shop"),
-                                              HomeThreeRowButton(
-                                                  iconURL:
-                                                      "assets/icons/home-calendar.svg",
-                                                  description: "Calendar"),
-                                            ],
-                                          ),
-                                        ),
-                                      ),
-                                      const SizedBox(
-                                        height: 30,
-                                      ),
-                                      SizedBox(
-                                        height: 25,
-                                        child: Container(
-                                          margin: EdgeInsets.only(
-                                              left:
-                                                  deviceData.size.width * 0.05),
-                                          child: Row(
-                                            children: [
-                                              Text(
-                                                "Latest Challenges",
-                                                style: GoogleFonts.almarai(
-                                                  color:
-                                                      const Color(0xFF1E1E1E),
-                                                  fontSize: 18,
-                                                  fontWeight: FontWeight.w700,
-                                                ),
-                                              ),
-                                            ],
-                                          ),
-                                        ),
-                                      ),
-                                      const SizedBox(
-                                        height: 5,
-                                      ),
-                                      // * Latest Challenges
-
-                                      Expanded(
-                                        child: ListView.builder(
-                                          itemCount: 10,
-                                          itemBuilder: (BuildContext context,
-                                              int index) {
-                                            return const LatestChallengeCard();
-                                          },
-                                        ),
-                                      )
-                                    ],
-                                  ),
-                                )
-                              ],
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
+        bottomNavigationBar: BottomNavigationBar(
+          showSelectedLabels: false,
+          showUnselectedLabels: false,
+          type: BottomNavigationBarType.fixed,
+          items: <BottomNavigationBarItem>[
+            BottomNavigationBarItem(
+              icon: SvgPicture.asset("assets/icons/challenges.svg"),
+              activeIcon: SvgPicture.asset(
+                "assets/icons/challenges.svg",
+                colorFilter:
+                    const ColorFilter.mode(Colors.black, BlendMode.srcIn),
               ),
+              label: 'Home',
             ),
-          )
-        ],
-      ),
-    );
+            BottomNavigationBarItem(
+              icon: SvgPicture.asset("assets/icons/search-low-opacity.svg"),
+              activeIcon: SvgPicture.asset(
+                "assets/icons/search-low-opacity.svg",
+                colorFilter:
+                    const ColorFilter.mode(Colors.black, BlendMode.srcIn),
+              ),
+              label: 'Search',
+            ),
+            BottomNavigationBarItem(
+              icon: SvgPicture.asset("assets/icons/home.svg"),
+              activeIcon: SvgPicture.asset(
+                "assets/icons/home.svg",
+                colorFilter:
+                    const ColorFilter.mode(Colors.black, BlendMode.srcIn),
+              ),
+              label: 'Home',
+            ),
+            BottomNavigationBarItem(
+              icon: SvgPicture.asset("assets/icons/community.svg"),
+              activeIcon: SvgPicture.asset(
+                "assets/icons/community.svg",
+                colorFilter:
+                    const ColorFilter.mode(Colors.black, BlendMode.srcIn),
+              ),
+              label: 'Community',
+            ),
+            BottomNavigationBarItem(
+              icon: SvgPicture.asset("assets/icons/notifications.svg"),
+              activeIcon: SvgPicture.asset(
+                "assets/icons/notifications.svg",
+                colorFilter:
+                    const ColorFilter.mode(Colors.black, BlendMode.srcIn),
+              ),
+              label: 'Notifications',
+            ),
+          ],
+          currentIndex: _selectedIndex,
+          onTap: _onItemTapped,
+        ),
+        resizeToAvoidBottomInset: false,
+        body: _widgetOptions.elementAt(_selectedIndex));
   }
 }
